## Resolving merge conflicts - Moved CassandraSnappyConnectionJob under `jobs` package which contains other jobs - Using `SnappyJobTestSupport` for `CassandraSnappyDUnitTest` for utility methods related to snappy job management
